pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docvault"
version = "0.1.0"
description = "A production-quality RAG system for querying PDF documents with source citations"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Traelyn", email = "traelyn.brasseaux@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "faiss-cpu>=1.9.0",
    "gradio>=4.44.0",
    "pymupdf>=1.24.0",
    "sentence-transformers>=3.3.0",
    "torch>=2.4.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
docvault-ingest = "scripts.ingest_docs:main"
docvault-query = "scripts.query:main"

[tool.hatch.build.targets.wheel]
packages = ["src/docvault"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.envs.default]
python = "3.10"
